+++ src/editor/dataset/constant/Element.ts
@@ -1,5 +1,6 @@
 import { ElementType } from '../enum/Element'
 import { IElement } from '../../interface/Element'
+import { ITd } from '../../interface/table/Td'
 
 export const EDITOR_ELEMENT_STYLE_ATTR: Array<keyof IElement> = [
   'bold',
@@ -59,6 +60,12 @@ export const EDITOR_ELEMENT_ZIP_ATTR: Array<keyof IElement> = [
   'listStyle',
   'listWrap',
   'groupIds'
+]
+
+export const TABLE_TD_ZIP_ATTR: Array<keyof ITd> = [
+  'verticalAlign',
+  'backgroundColor',
+  'borderType'
 ]
 
 export const TABLE_CONTEXT_ATTR: Array<keyof IElement> = [