+++ src/editor/dataset/constant/Regular.ts
@@ -1,8 +1,6 @@
 export const NUMBER_REG = /[0-9]/
-export const LETTER_REG = /[a-zA-Z]/
 export const NUMBER_LIKE_REG = /[0-9.]/
 export const CHINESE_REG = /[\u4e00-\u9fa5]/
-export const WORD_LIKE_REG = /[^a-zA-Z][a-zA-Z]/
 export const SURROGATE_PAIR_REG = /[\uD800-\uDBFF][\uDC00-\uDFFF]/ // unicode代理对（surrogate pair）
 // https://github.com/mathiasbynens/emoji-test-regex-pattern
 export const EMOJI_REG =