+++ src/editor/index.ts
@@ -44,7 +44,7 @@ import { defaultCursorOption } from './dataset/constant/Cursor'
 import { IPageNumber } from './interface/PageNumber'
 import { defaultPageNumberOption } from './dataset/constant/PageNumber'
 import { VerticalAlign } from './dataset/enum/VerticalAlign'
-import { TableBorder } from './dataset/enum/table/Table'
+import { TableBorder, TdBorder } from './dataset/enum/table/Table'
 import { IFooter } from './interface/Footer'
 import { defaultFooterOption } from './dataset/constant/Footer'
 import { MaxHeightRatio, NumberType } from './dataset/enum/Common'
@@ -63,6 +63,9 @@ import { IGroup } from './interface/Group'
 import { defaultGroupOption } from './dataset/constant/Group'
 import { IRangeStyle } from './interface/Listener'
 import { Override } from './core/override/Override'
+import { defaultPageBreakOption } from './dataset/constant/PageBreak'
+import { IPageBreak } from './interface/PageBreak'
+import { LETTER_CLASS } from './dataset/constant/Common'
 
 export default class Editor {
   public command: Command
@@ -118,6 +121,10 @@ export default class Editor {
       ...defaultGroupOption,
       ...options.group
     }
+    const pageBreakOptions: Required<IPageBreak> = {
+      ...defaultPageBreakOption,
+      ...options.pageBreak
+    }
 
     const editorOptions: DeepRequired<IEditorOption> = {
       mode: EditorMode.EDIT,
@@ -148,8 +155,8 @@ export default class Editor {
       marginIndicatorColor: '#BABABA',
       margins: [100, 120, 100, 120],
       pageMode: PageMode.PAGING,
-      tdPadding: 5,
-      defaultTrMinHeight: 40,
+      tdPadding: [0, 5, 5, 5],
+      defaultTrMinHeight: 42,
       defaultColMinWidth: 40,
       defaultHyperlinkColor: '#0000FF',
       paperDirection: PaperDirection.VERTICAL,
@@ -158,6 +165,7 @@ export default class Editor {
       wordBreak: WordBreak.BREAK_WORD,
       printPixelRatio: 3,
       maskMargin: [0, 0, 0, 0],
+      letterClass: [LETTER_CLASS.ENGLISH],
       ...options,
       header: headerOptions,
       footer: footerOptions,
@@ -168,7 +176,8 @@ export default class Editor {
       cursor: cursorOptions,
       title: titleOptions,
       placeholder: placeholderOptions,
-      group: groupOptions
+      group: groupOptions,
+      pageBreak: pageBreakOptions
     }
     // 数据处理
     let headerElementList: IElement[] = []
@@ -236,6 +245,8 @@ export default class Editor {
 
 // 对外对象
 export {
+  EDITOR_COMPONENT,
+  LETTER_CLASS,
   Editor,
   RowFlex,
   VerticalAlign,
@@ -244,7 +255,6 @@ export {
   ElementType,
   ControlType,
   EditorComponent,
-  EDITOR_COMPONENT,
   PageMode,
   ImageDisplay,
   Command,
@@ -252,6 +262,7 @@ export {
   BlockType,
   PaperDirection,
   TableBorder,
+  TdBorder,
   MaxHeightRatio,
   NumberType,
   TitleLevel,